requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
pyttsx3
nltk
Levenshtein
huggingface-hub==0.14.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2